// generated/Marketplace-Replanted/Beanstalk.ts
export class PodListingCreated extends ethereum.Event {
  get params(): PodListingCreated__Params {
    return new PodListingCreated__Params(this);
  }
}

export class PodListingCreated__Params {
  _event: PodListingCreated;

  constructor(event: PodListingCreated) {
    this._event = event;
  }

  get account(): Address {
    return this._event.parameters[0].value.toAddress();
  }

  get index(): BigInt {
    return this._event.parameters[1].value.toBigInt();
  }

  get start(): BigInt {
    return this._event.parameters[2].value.toBigInt();
  }

  get amount(): BigInt {
    return this._event.parameters[3].value.toBigInt();
  }

  get pricePerPod(): i32 {
    return this._event.parameters[4].value.toI32();
  }

  get maxHarvestableIndex(): BigInt {
    return this._event.parameters[5].value.toBigInt();
  }

  get mode(): i32 {
    return this._event.parameters[6].value.toI32();
  }
}